requirements.txt
# Healthcare Policy Assistant - Production Requirements
# Streamlit Cloud + Databricks SQL Warehouse

# Core Framework
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0

# Databricks SQL Connection (REQUIRED for SQL Warehouse)
databricks-sql-connector>=3.0.0

# LLM & AI
openai>=1.12.0

# Graph & Visualization
networkx>=3.2
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
